refactor: rename tracepath module to graph
The functionality implemented in there is generic enough that we can use
it as a general graph module.
