+++ src/debsbom/graph/walker.py
